requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
PySocks==1.7.1